requirements.txt
asgiref==3.4.1
Django==3.2.7
djangorestframework==3.12.4
numpy==1.21.2
Pillow==8.3.2
pytz==2021.1
sqlparse==0.4.2
torch==1.9.1
torchaudio==0.9.1
torchvision==0.10.1
typing-extensions==3.10.0.2
